backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0

pydantic==2.5.0
pydantic-settings==2.1.0

# Authentication
python-jose[cryptography]
passlib[bcrypt]
python-multipart

aioboto3==11.3.1

python-dotenv==1.0.0

# Resilience and reliability
tenacity>=8.2.3
circuitbreaker>=2.0.0
aiocache>=0.12.2
